registration tests + some clean up
