+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 -   repo: https://github.com/psf/black
-    rev: 22.6.0
+    rev: 25.1.0
     hooks:
     -   id: black
 -   repo: git@github.com:Yelp/detect-secrets
