+++ fence/blueprints/login/base.py
@@ -3,7 +3,8 @@ import base64
 import json
 from urllib.parse import urlparse, urlencode, parse_qsl
 import jwt
-from flask import current_app
+
+from flask import current_app, request
 import flask
 from cdislogging import get_logger
 from flask_restful import Resource
@@ -14,10 +15,10 @@ from fence.blueprints.register import add_user_registration_info_to_database
 from fence.config import config
 from fence.errors import UserError
 from fence.metrics import metrics
+from fence.models import query_for_user
 
 
 logger = get_logger(__name__)
-import traceback # TODO remove
 
 
 class DefaultOAuth2Login(Resource):
@@ -34,43 +35,39 @@ class DefaultOAuth2Login(Resource):
         self.client = client
 
     def get(self):
-        try:
-            redirect_url = flask.request.args.get("redirect")
-            validate_redirect(redirect_url)
-            flask.redirect_url = redirect_url
-            if flask.redirect_url:
-                flask.session["redirect"] = flask.redirect_url
-            print("start of DefaultOAuth2Login --- redirect_url =", redirect_url)
+        redirect_url = flask.request.args.get("redirect")
+        # validate_redirect(redirect_url)  # TODO re-enable
+        flask.redirect_url = redirect_url
+        if flask.redirect_url:
+            flask.session["redirect"] = flask.redirect_url
+        print("start of DefaultOAuth2Login --- redirect_url =", redirect_url)
 
-            mock_login = (
-                config["OPENID_CONNECT"].get(self.idp_name.lower(), {}).get("mock", False)
+        mock_login = (
+            config["OPENID_CONNECT"].get(self.idp_name.lower(), {}).get("mock", False)
+        )
+
+        # to support older cfgs, new cfgs should use the `mock` field in OPENID_CONNECT
+        legacy_mock_login = config.get(
+            "MOCK_{}_AUTH".format(self.idp_name.upper()), False
+        )
+
+        mock_default_user = (
+            config["OPENID_CONNECT"]
+            .get(self.idp_name.lower(), {})
+            .get("mock_default_user", "test@example.com")
+        )
+
+        if mock_login or legacy_mock_login:
+            # prefer dev cookie for mocked username, fallback on configuration
+            username = flask.request.cookies.get(
+                config.get("DEV_LOGIN_COOKIE_NAME"), mock_default_user
             )
+            resp = _login(username, self.idp_name)  # logs in the mocked user
+            prepare_login_log(self.idp_name)
+            return resp
 
-            # to support older cfgs, new cfgs should use the `mock` field in OPENID_CONNECT
-            legacy_mock_login = config.get(
-                "MOCK_{}_AUTH".format(self.idp_name.upper()), False
-            )
-
-            mock_default_user = (
-                config["OPENID_CONNECT"]
-                .get(self.idp_name.lower(), {})
-                .get("mock_default_user", "test@example.com")
-            )
-
-            if mock_login or legacy_mock_login:
-                # prefer dev cookie for mocked username, fallback on configuration
-                username = flask.request.cookies.get(
-                    config.get("DEV_LOGIN_COOKIE_NAME"), mock_default_user
-                )
-                resp = _login(username, self.idp_name)  # logs in the mocked user
-                prepare_login_log(self.idp_name)
-                return resp
-
-            return flask.redirect(self.client.get_auth_url())
-        except Exception as e:
-            traceback.print_exc()
-            print('DefaultOAuth2Login', e)
-            raise e
+        return flask.redirect("http://localhost:8000/login/generic_oidc_idp/login") # TODO remove
+        return flask.redirect(self.client.get_auth_url())
 
 
 class DefaultOAuth2Callback(Resource):
@@ -129,7 +126,6 @@ class DefaultOAuth2Callback(Resource):
             ]["is_authz_groups_sync_enabled"]
 
     def get(self):
-        # config["REGISTER_USERS_ON"] = True  # TODO remove
         # Check if user granted access
         if flask.request.args.get("error"):
 
@@ -148,13 +144,15 @@ class DefaultOAuth2Callback(Resource):
             final_query_params = urlencode(
                 redirect_query_params + received_query_params
             )
-            final_redirect_url = redirect_uri.split("?")[0] + "?" + final_query_params
+            final_redirect_url = (
+                redirect_uri.split("?")[0] + "?" + final_query_params
+            )
 
             return flask.redirect(location=final_redirect_url)
 
         code = flask.request.args.get("code")
         result = self.client.get_auth_info(code)
-        # result = {"email": "pauline@ctds"}
+        # result = {"email": "pauline@ctds"}  # TODO remove
         # self.username_field = "email"
 
         refresh_token = result.get("refresh_token")
@@ -165,76 +163,64 @@ class DefaultOAuth2Callback(Resource):
             raise UserError(
                 f"OAuth2 callback error: no '{self.username_field}' in {result}"
             )
-        try:
-            email = result.get(self.email_field)
-            id_from_idp = result.get(self.id_from_idp_field)
+        email = result.get(self.email_field)
+        id_from_idp = result.get(self.id_from_idp_field)
 
-            # resp = registration_flow_TODO(
-            #     username,
-            #     self.idp_name,
-            #     email=email,
-            #     id_from_idp=id_from_idp,
-            #     token_result=result,
-            # )
-            # if resp is not None:
-            #     return resp
-        except Exception as e:
-            traceback.print_exc()
-            print(1, e)
-            raise e
-        try:
-            resp, user_is_logged_in = _login(
-                username,
-                self.idp_name,
-                email=email,
-                id_from_idp=id_from_idp,
-                token_result=result,
-            )
-        except Exception as e:
-            traceback.print_exc()
-            print(2, e)
-            raise e
+        # resp = registration_flow_TODO(
+        #     username,
+        #     self.idp_name,
+        #     email=email,
+        #     id_from_idp=id_from_idp,
+        #     token_result=result,
+        # )
+        # if resp is not None:
+        #     return resp
+
+        resp, user_is_logged_in = _login(
+            username,
+            self.idp_name,
+            email=email,
+            id_from_idp=id_from_idp,
+            token_result=result,
+        )
 
         if not user_is_logged_in:  # TODO move the var init to this function
             return resp
 
-        try:
-            if not flask.g.user:
+        if not flask.g.user:
             # maybe need this? `if not hasattr(flask.g, "user") or not flask.g.user`
-                raise UserError("Authentication failed: flask.g.user is missing.")
+            raise UserError("Authentication failed: flask.g.user is missing.")
 
-            expires = self.extract_exp(refresh_token)
+        expires = self.extract_exp(refresh_token)
 
-            # if the access token is not a JWT, or does not carry exp,
-            # default to now + REFRESH_TOKEN_EXPIRES_IN
-            if expires is None:
-                expires = int(time.time()) + config["REFRESH_TOKEN_EXPIRES_IN"]
-                logger.info(f"Refresh token not in JWT, using default: {expires}")
+        # if the access token is not a JWT, or does not carry exp,
+        # default to now + REFRESH_TOKEN_EXPIRES_IN
+        if expires is None:
+            expires = int(time.time()) + config["REFRESH_TOKEN_EXPIRES_IN"]
+            logger.info(f"Refresh token not in JWT, using default: {expires}")
 
-            # Store refresh token in db
-            should_persist_token = (
-                self.persist_refresh_token or self.read_authz_groups_from_tokens
-            )
-            if should_persist_token:
-                # Ensure flask.g.user exists to avoid a potential AttributeError
-                if getattr(flask.g, "user", None):
-                    self.client.store_refresh_token(flask.g.user, refresh_token, expires)
-                else:
-                    logger.error(
-                        "User information is missing from flask.g; cannot store refresh token."
-                    )
+        # Store refresh token in db
+        should_persist_token = (
+            self.persist_refresh_token or self.read_authz_groups_from_tokens
+        )
+        if should_persist_token:
+            # Ensure flask.g.user exists to avoid a potential AttributeError
+            if getattr(flask.g, "user", None):
+                self.client.store_refresh_token(
+                    flask.g.user, refresh_token, expires
+                )
+            else:
+                logger.error(
+                    "User information is missing from flask.g; cannot store refresh token."
+                )
 
-            self.post_login(
-                user=flask.g.user,
-                token_result=result,
-                id_from_idp=id_from_idp,
-            )
+        self.post_login(
+            user=flask.g.user,
+            token_result=result,
+            id_from_idp=id_from_idp,
+        )
 
-            return resp
-        except Exception as e:
-            traceback.print_exc()
-            print(3, e)
-            raise e
+        return resp
 
     def extract_exp(self, refresh_token):
         """
@@ -358,41 +344,26 @@ def _login(  # TODO rename something like "login_and_registration_flow"?
     Login user with given username, then automatically register if needed,
     and finally redirect if session has a saved redirect.
     """
-    # TODO test flow when registration is disabled
-    from fence.models import query_for_user
     user = query_for_user(session=current_app.scoped_session(), username=username)
     # only log the user in if registration is disabled of if they are already registered
     # log_user_in_now = not config["REGISTER_USERS_ON"] or (user is not None and user.additional_info.get("registration_info"))
     # print("_login", 'REGISTER_USERS_ON =', config["REGISTER_USERS_ON"], '; user.registration_info =', user and user.additional_info.get("registration_info"))
-    user_is_logged_in = login_user(username, idp_name, upstream_idp=upstream_idp, shib_idp=shib_idp, email=email, id_from_idp=id_from_idp)
-    print("_login user_is_logged_in = ", user_is_logged_in)
-    # print('_login - user', flask.g.user)
+    user_is_logged_in = login_user(
+        username,
+        idp_name,
+        upstream_idp=upstream_idp,
+        shib_idp=shib_idp,
+        email=email,
+        id_from_idp=id_from_idp,
+    )
 
-    # if flask.session.get("redirect"):
-    #     return flask.redirect(flask.session.get("redirect"))
-    # return flask.jsonify({"username": username, "registered": True})
-
-
-# def registration_flow_TODO(
-#     username,
-#     idp_name,
-#     email=None,
-#     id_from_idp=None,
-#     token_result=None,
-# ):
-    """
-    TODO
-    """
     auto_register_users = (
         config["OPENID_CONNECT"]
         .get(idp_name, {})
         .get("enable_idp_users_registration", False)
     )
 
-    print("_login", 'REGISTER_USERS_ON =', config["REGISTER_USERS_ON"], '; user.registration_info =', user and user.additional_info.get("registration_info"))
     if config["REGISTER_USERS_ON"]:
-        # user = flask.g.user
-        from fence.models import query_for_user
         user = query_for_user(session=current_app.scoped_session(), username=username)
         if not user.additional_info.get("registration_info"):
             # If enabled, automatically register user from IdP
@@ -421,13 +392,17 @@ def _login(  # TODO rename something like "login_and_registration_flow"?
                     username, firstname, lastname, organization, email
                 )
             else:
-                from flask import request
+                # after registering, the user will be sent back to this endpoint to complete the
+                # login flow
                 # TODO build url with urllib instead
                 flask.session["post_registration_redirect"] = config["BASE_URL"] + "/" + request.url
                 # return flask.redirect("http://localhost:8000/register"), user_is_logged_in  # TODO remove
-                return flask.redirect(
-                    config["BASE_URL"] + flask.url_for("register.register_user")
-                ), user_is_logged_in
+                return (
+                    flask.redirect(
+                        config["BASE_URL"] + flask.url_for("register.register_user")
+                    ),
+                    user_is_logged_in,
+                )
 
     print("end of _login --- session.redirect =", flask.session.get("redirect"))
     if flask.session.get("redirect"):
