+++ tests/test_metrics.py
@@ -23,13 +23,11 @@ tests looking at users are not affected.
 """
 
 import boto3
-import flask
 import json
 import jwt
 import mock
 import pytest
-import time
-from unittest.mock import ANY, MagicMock, patch
+from unittest.mock import MagicMock, patch
 
 import fence
 from fence.blueprints.login import get_idp_route_name
@@ -453,9 +451,8 @@ def test_presigned_url_log_unauthorized(
 def test_login_log_login_endpoint(
     client,
     idp,
-    mock_arborist_requests,
-    rsa_private_key,
     db_session,  # do not remove :-) See note at top of file
+    mocks_for_idp_oauth2_callbacks,
     monkeypatch,
     prometheus_metrics_before,
 ):
@@ -463,86 +460,12 @@ def test_login_log_login_endpoint(
     Test that logging in via any of the existing IDPs triggers the creation
     of a login audit log and of a prometheus metric.
     """
-    mock_arborist_requests()
     audit_service_mocker = mock.patch(
         "fence.resources.audit.client.requests", new_callable=mock.Mock
     )
     monkeypatch.setitem(config, "ENABLE_AUDIT_LOGS", {"login": True})
 
-    username = "test@test"
-    callback_endpoint = "login"
-    idp_name = idp
-    headers = {}
-    get_auth_info_value = {}
-    jwt_string = jwt.encode(
-        {"iat": int(time.time())}, key=rsa_private_key, algorithm="RS256"
-    )
-
-    if idp == "synapse":
-        get_auth_info_value = {
-            "fence_username": username,
-            "sub": username,
-            "given_name": username,
-            "family_name": username,
-        }
-    elif idp == "orcid":
-        get_auth_info_value = {"orcid": username}
-    elif idp == "cilogon":
-        get_auth_info_value = {"sub": username}
-    elif idp == "shibboleth":
-        headers["persistent_id"] = username
-        idp_name = "itrust"
-    elif idp == "okta":
-        get_auth_info_value = {"okta": username}
-    elif idp == "fence":
-        mocked_fetch_access_token = MagicMock(return_value={"id_token": jwt_string})
-        patch(
-            f"authlib.integrations.flask_client.apps.FlaskOAuth2App.fetch_access_token",
-            mocked_fetch_access_token,
-        ).start()
-        mocked_validate_jwt = MagicMock(
-            return_value={"context": {"user": {"name": username}}}
-        )
-        patch(
-            f"fence.blueprints.login.fence_login.validate_jwt", mocked_validate_jwt
-        ).start()
-    elif idp == "ras":
-        get_auth_info_value = {"username": username}
-        callback_endpoint = "callback"
-        # these should be populated by a /login/<idp> call that we're skipping:
-        flask.g.userinfo = {"sub": "testSub123"}
-        flask.g.tokens = {
-            "refresh_token": jwt_string,
-            "id_token": jwt_string,
-        }
-        flask.g.encoded_visas = ""
-    elif idp == "generic_with_discovery_url":
-        get_auth_info_value = {"generic_with_discovery_url_username": username}
-    elif idp == "generic_additional_params":
-        # get_auth_info_value specific to generic_additional_params
-        # TODO: Need test when is_authz_groups_sync_enabled == true
-        get_auth_info_value = {
-            "username": username,
-            "sub": username,
-            "email_verified": True,
-            "iat": 1609459200,
-            "exp": 1609462800,
-            "refresh_token": "mock_refresh_token",
-            "groups": ["group1", "group2"],
-        }
-    elif idp.startswith("generic_"):
-        get_auth_info_value = {"sub": username}
-
-    if idp in ["google", "microsoft", "okta", "synapse", "cognito"]:
-        get_auth_info_value["email"] = username
-
-    get_auth_info_patch = None
-    if get_auth_info_value:
-        mocked_get_auth_info = MagicMock(return_value=get_auth_info_value)
-        get_auth_info_patch = patch(
-            f"flask.current_app.{idp}_client.get_auth_info", mocked_get_auth_info
-        )
-        get_auth_info_patch.start()
+    idp_name, username, callback_endpoint, headers = mocks_for_idp_oauth2_callbacks
 
     with audit_service_mocker as audit_service_requests:
         audit_service_requests.post.return_value = MockResponse(
@@ -567,9 +490,6 @@ def test_login_log_login_endpoint(
                 "client_id": None,
             },
         )
-
-    if get_auth_info_patch:
-        get_auth_info_patch.stop()
 
     # check prometheus metrics
     resp = client.get("/metrics")
